requirements.txt
google-api-python-client==2.48.0
google-auth==2.11.0
google-auth-oauthlib==0.5.1
google-auth-httplib2==0.1.0
requests==2.28.1
flashtext==2.7
